+++ test_steps/__init__.py
@@ -8,7 +8,7 @@ like py.test or nose
 """
 
 __author__ = 'Steven LI'
-__version__ = '0.8.3'
+__version__ = '0.8.5'
 
 import logging
 import os, re, time