+++ CHANGES.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+v0.8.5
+-------------------------------------------
+* Fix #24: relative path handling in yaml_testbed function. Now it will check the path relative to the cwd(),
+if not exists, it will try to get the file relative to the parent file location.
+
 v0.8.3
 -------------------------------------------
 * Fix #25: default test bed file support. Now you can invoke init_testbed() without parameters.